requirements.txt
Django
djangorestframework
psycopg2
redis
djangorestframework
djangorestframework-simplejwt
django-rest-framework-social-oauth2
social-auth-app-django

